requirements.txt
django==5.1.1
django-heroku
python-decouple
channels
djangorestframework
